requirements.txt
pandas
SQLAlchemy
sqlalchemy-redshift
psycopg2-binary
python-dotenv
pendulum
httpx